pyproject.toml
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"    # ← change this line



[project]
name = "tiny-rag-ai"
version = "0.1.2"
readme = "README.md"
description = "Complete In-Device RAG chatbot library, no API keys needed.With 2 lines of implementation in applications."
requires-python = ">=3.9"

dependencies = [
    "faiss-cpu>=1.7.4",       
    "numpy>=1.24.0",
    "sentence-transformers>=2.7.0",
    "llama-cpp-python>=0.2.0",
    "huggingface-hub>=0.23.0",
    "pathlib>=1.0.1", 
]

[project.urls]
Homepage = "https://github.com/Madhan-1000/tiny-rag-ai"
Repository = "https://github.com/Madhan-1000/tiny-rag-ai"


[project.optional-dependencies]
gpu = [
    "faiss-gpu>=1.7.4",        
]

[tool.setuptools.packages.find]
where = ["src"]